+++ crates/types/src/lib.rs
@@ -1,5 +1,6 @@
 mod arrays;
 mod enums;
+mod externals;
 mod literals;
 mod numbers;
 mod objects;
@@ -10,6 +11,7 @@ mod unions;
 
 pub use arrays::*;
 pub use enums::*;
+pub use externals::*;
 pub use literals::*;
 pub use numbers::*;
 pub use objects::*;
@@ -75,7 +77,7 @@ impl SchemaType {
         })
     }
 
-    /// Create a nullable type for the provided schema. If already nullabe,
+    /// Convert the provided schema to a nullable type. If already nullable,
     /// do nothing and return, otherwise convert to a union.
     pub fn nullable(mut schema: SchemaType) -> SchemaType {
         if let SchemaType::Union(inner) = &mut schema {