+++ CHANGELOG.md
@@ -2,6 +2,19 @@
 
 ## Unreleased
 
+#### 💥 Breaking
+
+- Renamed `SettingPath` to `Path`.
+- Renamed `Segment` to `PathSegment`.
+
+#### 🚀 Updates
+
+- Generator
+  - Updated JSON schema arrays to use `contains` when applicable.
+- Schema
+  - Added support for `chrono` types (behind the `type_chrono` feature).
+  - Added support for `regex` types (behind the `type_regex` feature).
+
 #### 🐞 Fixes
 
 - Generator